pyproject.toml
[project]
name = "imp_bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
]
